artical/requirements.txt

streamlit
crewai
langchain-openai
langchain-groq
langchain-anthropic
python-dotenv
